--- src/components/Forms/BaseHelperText/BaseHelperText.types.ts
@@ -1,16 +1,0 @@
-/* Internal dependencies */
-import { BezierComponentProps, ChildrenProps, IdentifierProps } from 'Types/ComponentProps'
-import { FormComponentProps } from 'Components/Forms/Form.types'
-import { TextProps } from 'Components/Text'
-
-interface BaseHelperTextOptions {
-  type: 'info' | 'error'
-}
-
-export default interface BaseHelperTextProps extends
-  BezierComponentProps,
-  ChildrenProps,
-  Partial<IdentifierProps>,
-  TextProps,
-  Pick<FormComponentProps, 'hasError'>,
-  BaseHelperTextOptions {}
