--- src/components/Forms/BaseHelperText/index.ts
@@ -1,2 +1,0 @@
-export { default as BaseHelperText } from './BaseHelperText'
-export type { default as BaseHelperTextProps } from './BaseHelperText.types'
