+++ src/components/Forms/FormControl/FormControl.stories.tsx
@@ -13,8 +13,7 @@ import { FormLabel } from 'Components/Forms/FormLabel'
 import { TextField } from 'Components/Forms/Inputs/TextField'
 import { TextArea } from 'Components/Forms/Inputs/TextArea'
 import { Select } from 'Components/Forms/Inputs/Select'
-import { FormHelperText } from 'Components/Forms/FormHelperText'
-import { FormErrorMessage } from 'Components/Forms/FormErrorMessage'
+import { FormHelperText, FormErrorMessage } from 'Components/Forms/FormHelperText'
 import FormControl from './FormControl'
 import FormControlProps from './FormControl.types'
