--- src/components/Forms/FormErrorMessage/FormErrorMessage.types.ts
@@ -1,4 +1,0 @@
-/* Internal dependencies */
-import { BaseHelperTextProps } from 'Components/Forms/BaseHelperText'
-
-export default interface FormErrorMessageProps extends Omit<BaseHelperTextProps, 'type'> {}
