--- src/components/Forms/FormErrorMessage/index.ts
@@ -1,2 +1,0 @@
-export { default as FormErrorMessage } from './FormErrorMessage'
-export type { default as FormErrorMessageProps } from './FormErrorMessage.types'
