+++ src/components/Forms/FormHelperText/FormHelperText.test.tsx
@@ -4,8 +4,13 @@ import React from 'react'
 /* Internal dependencies */
 import { LightFoundation } from 'Foundation'
 import { render } from 'Utils/testUtils'
-import FormHelperText, { FORM_HELPER_TEXT_TEST_ID } from './FormHelperText'
-import type FormHelperTextProps from './FormHelperText.types'
+import {
+  FORM_HELPER_TEXT_TEST_ID,
+  FORM_ERROR_MESSAGE_TEST_ID,
+  FormHelperText,
+  FormErrorMessage,
+} from './FormHelperText'
+import type { FormHelperTextProps, FormErrorMessageProps } from './FormHelperText.types'
 
 describe('FormHelperText >', () => {
   let props: FormHelperTextProps
@@ -42,3 +47,47 @@ describe('FormHelperText >', () => {
     expect(helperText).toBeNull()
   })
 })
+
+describe('FormErrorMessage >', () => {
+  let props: FormErrorMessageProps
+  const text = 'Lorem ipsum'
+
+  beforeEach(() => {
+    props = {
+      id: 'test',
+      children: text,
+    }
+  })
+
+  const renderFormHelperText = (otherProps?: Partial<FormErrorMessageProps>) =>
+    render(<FormErrorMessage {...props} {...otherProps} />)
+
+  it('Snapshot >', () => {
+    const { getByTestId } = renderFormHelperText()
+    const rendered = getByTestId(FORM_ERROR_MESSAGE_TEST_ID)
+
+    expect(rendered).toMatchSnapshot()
+  })
+
+  it('renders text with correct style when children prop is not empty', () => {
+    const { getByTestId } = renderFormHelperText()
+    const rendered = getByTestId(FORM_ERROR_MESSAGE_TEST_ID)
+
+    expect(rendered).toHaveStyle(`color: ${LightFoundation.theme['bgtxt-orange-normal']}`)
+  })
+
+  it('renders nothing when children prop is empty', () => {
+    const { queryByTestId } = renderFormHelperText({ children: '' })
+    const rendered = queryByTestId(FORM_ERROR_MESSAGE_TEST_ID)
+
+    expect(rendered).toBeNull()
+  })
+
+  it('should have aria-live="polite" attribute', () => {
+    const { getByTestId } = renderFormHelperText()
+    const rendered = getByTestId(FORM_ERROR_MESSAGE_TEST_ID)
+
+    expect(rendered).toHaveAttribute('aria-live', 'polite')
+  })
+})
+
