+++ src/components/Forms/FormHelperText/FormHelperText.types.ts
@@ -1,4 +1,20 @@
 /* Internal dependencies */
-import { BaseHelperTextProps } from 'Components/Forms/BaseHelperText'
+import { BezierComponentProps, ChildrenProps, IdentifierProps } from 'Types/ComponentProps'
+import { FormComponentProps } from 'Components/Forms/Form.types'
+import { TextProps } from 'Components/Text'
 
-export default interface FormHelperTextProps extends Omit<BaseHelperTextProps, 'type'> {}
+interface BaseHelperTextOptions {
+  type: 'info' | 'error'
+}
+
+export interface BaseHelperTextProps extends
+  BezierComponentProps,
+  ChildrenProps,
+  Partial<IdentifierProps>,
+  TextProps,
+  Pick<FormComponentProps, 'hasError'>,
+  BaseHelperTextOptions {}
+
+export interface FormHelperTextProps extends Omit<BaseHelperTextProps, 'type'> {}
+
+export interface FormErrorMessageProps extends Omit<BaseHelperTextProps, 'type'> {}
